Moved variables from inside class to global scope - works now
